Adds auth to all task methods

+++ backend-nestjs/src/tasks/tasks.service.ts
@@ -14,12 +14,14 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(getFilterDto: GetTasksFilterDto) {
-    return this.taskRepository.getTasks(getFilterDto);
+  async getTasks(getFilterDto: GetTasksFilterDto, user: User) {
+    return this.taskRepository.getTasks(getFilterDto, user);
   }
 
-  async getTaskById(id: number) {
-    const task = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user) {
+    const task = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!task) {
       throw new NotFoundException();
     }
@@ -30,16 +32,20 @@ export class TasksService {
     return await this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
